Erste implementierung einer queued DeviceAsync. PacketTypes wurde refaktored.
git-svn-id: https://svn.itm.uni-luebeck.de/fallstudie2010/sources@37 0a1f239c-e987-4cba-9274-3ef986406e12

+++ tr.wsn-device-drivers/src/main/java/de/uniluebeck/itm/devicedriver/async/OperationQueue.java
@@ -20,7 +20,7 @@ public interface OperationQueue {
 	 * @param callback The callback interface that is called when the operation is done.
 	 * @return A <code>OperationHandle</code> for controlling the operation.
 	 */
-	<T> OperationHandle<T> addOperation(Operation<T> operation, int timeout, AsyncCallback<T> callback);
+	<T> OperationHandle<T> addOperation(Operation<T> operation, long timeout, AsyncCallback<T> callback);
 	
 	/**
 	 * Returns all operations that are in the queue.
